backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
celery==5.3.6
redis==5.0.1
requests==2.31.0
httpx==0.26.0
pandas==2.2.0
numpy==1.26.3
scikit-learn==1.4.0
tensorflow==2.15.0
openai==1.10.0
langchain==0.1.4
langchain-openai==0.0.5
pypdf2==3.0.1
openpyxl==3.1.2
python-dateutil==2.8.2
pytz==2024.1
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0
faker==22.6.0
